+++ web/src/components/post-item.vue
@@ -143,6 +143,10 @@ import {
     BookmarkOutline,
     ChatboxOutline,
     ShareSocialOutline,
+    PersonAddOutline,
+    PersonRemoveOutline,
+    BodyOutline,
+    WalkOutline,
 } from '@vicons/ionicons5';
 import { MoreHorizFilled } from '@vicons/material';
 import copy from "copy-to-clipboard";
@@ -151,10 +155,14 @@ const router = useRouter();
 const store = useStore();
 const props = withDefaults(defineProps<{
     post: Item.PostProps,
+    isOwner: boolean,
+    addExtraAction: boolean,
 }>(), {});
 
 const emit = defineEmits<{
-    (e: 'send-whisper', user: Item.UserInfo): void;
+    (e: 'send-whisper', user: Item.UserInfo): void
+    (e: 'handle-follow-action', user: Item.PostProps): void
+    (e: 'handle-friend-action', user: Item.PostProps): void
 }>();
 
 const renderIcon = (icon: Component) => {
@@ -180,6 +188,34 @@ const tweetOptions = computed(() => {
         key: 'whisper',
         icon: renderIcon(PaperPlaneOutline)
     });
+    if (!props.isOwner && props.addExtraAction) {
+        if (props.post.user.is_following) {
+            options.push({
+                label: '取消关注',
+                key: 'unfollow',
+                icon: renderIcon(WalkOutline)
+            })
+        } else {
+            options.push({
+                label: '关注',
+                key: 'follow',
+                icon: renderIcon(BodyOutline)
+            })
+        }
+        // if (props.post.user.is_friend) {
+        //     options.push({
+        //         label: '删除好友',
+        //         key: 'delete',
+        //         icon: renderIcon(PersonRemoveOutline)
+        //     });
+        // } else {
+        //     options.push({
+        //         label: '添加朋友',
+        //         key: 'requesting',
+        //         icon: renderIcon(PersonAddOutline)
+        //     });
+        // }
+    }
     options.push({
         label: '复制链接',
         key: 'copyTweetLink',
@@ -189,7 +225,7 @@ const tweetOptions = computed(() => {
 });
 
 const handleTweetAction = async (
-    item: 'copyTweetLink' | 'whisper'
+    item: 'copyTweetLink' | 'whisper' | 'follow' | 'unfollow' | 'delete' | 'requesting'
 ) => {
     switch (item) {
         case 'copyTweetLink':
@@ -198,6 +234,14 @@ const handleTweetAction = async (
             break;
         case 'whisper':
             emit('send-whisper', props.post.user);
+            break;
+        case 'delete':
+        case 'requesting':
+            emit('handle-friend-action', props.post);
+            break;
+        case 'follow':
+        case 'unfollow':
+            emit('handle-follow-action', props.post);
             break;
         default:
             break;